Cast `size_t' to `unsigned int'. Needed by Mac OS X.
This cast has been missed by the prefious fix because it was in the debugging
code which was disabled then..

+++ src/liboping.c
@@ -370,7 +370,7 @@ static int ping_receive_one (int fd, pinghost_t *ph, struct timeval *now)
 		return (-1);
 	}
 
-	dprintf ("Read %i bytes from fd = %i\n", buffer_len, fd);
+	dprintf ("Read %u bytes from fd = %i\n", (unsigned int) buffer_len, fd);
 
 	if (sa.ss_family == AF_INET)
 	{
